Minor various fixes

+++ svalbardradar/interpretations.py
@@ -375,7 +375,7 @@ def merge_all_interpretations(step_m: float = 5.0, overwrite_cache: bool = False
 
         out[f"{prefix}_user_std"] = grouped["depth"].std()
         out[f"{prefix}_user_nmad"] = grouped["depth"].apply(lambda v: statistics.nmad(v))
-        out[f"{prefix}_user_count"] = grouped["depth"].count()
+        out[f"{prefix}_user_count"] = grouped["depth"].count().astype(int)
 
         out[f"{prefix}_gpr_uncertainty"] = gpr_uncertainty(
             thickness=out[prefix], frequency_mhz=out["antenna"].str.replace(" MHz", "").astype(float)
